refactor: trail creation show loading spinner
