+++ src/components/Layout/Navbar/index.tsx
@@ -11,7 +11,7 @@ import {
   Nav, MobileMenuIcon, ResponsiveNavbar, Menu,
 } from './styles';
 
-const STRIPE_PRODUCT_URL = 'https://buy.stripe.com/test_5kAdQYdU059Des06oo';
+const STRIPE_PRODUCT_URL = 'https://buy.stripe.com/test_6oE5ksaHO1Xr6Zy9AB';
 
 const Navbar = () => {
   const [showMenu, setShowMenu] = useState(false);
